requirements.txt
# requirements.txt

requests
GitPython